requirements.txt
keras
tensorflow-gpu
keras-self-attention
h5py
obspy
pandas
